Changed Integer to BigDecimal. close #16

+++ src/main/groovy/br/com/mlm/dostock/domain/ProductBatch.groovy
@@ -8,7 +8,7 @@ class ProductBatch extends BaseEntity{
     @Column(nullable = false)
     String number
 
-    Integer quantity = 0
+    BigDecimal quantity = 0
 
     @Temporal(TemporalType.TIMESTAMP)
     Date expirationDate
