+++ src/main/groovy/br/com/mlm/dostock/services/impl/ProductLogServiceImpl.groovy
@@ -24,7 +24,7 @@ class ProductLogServiceImpl implements ProductLogService {
     }
 
     @Override
-    void register(Product product, Folder folder, ProductBatch productBatch, Integer quantity, String observation, ProductLogType logType) {
+    void register(Product product, Folder folder, ProductBatch productBatch, BigDecimal quantity, String observation, ProductLogType logType) {
         ProductLog productLog = new ProductLog()
         productLog.quantity = quantity
         productLog.observation = observation
